[#28] feature : Modal 공통 컴포넌트 구현
