+++ imgoing-frontend/src/components/Modal/MenuModal.tsx
@@ -38,28 +38,22 @@ const ModalButton = styled.TouchableOpacity<Pick<ModalButtonProps, 'first'>>`
   }}
 `;
 
-const MenuModal = (props: ModalProps) => {
-  const { isModalVisible, setIsModalVisible, isDeleteModalVisible, setIsDeleteModalVisible } =
-    props;
-
+const MenuModal = () => {
   return (
-    <Modal animationType={'slide'} transparent={true} visible={isModalVisible}>
-      <ModalView onPress={() => setIsModalVisible(!isModalVisible)}>
-        <ModalButton first onPress={() => setIsModalVisible(!isModalVisible)}>
-          <CalloutTypo color={'black'}>고정하기</CalloutTypo>
-        </ModalButton>
-        <ModalButton onPress={() => setIsModalVisible(!isModalVisible)}>
-          <CalloutTypo color={'black'}>편집하기</CalloutTypo>
-        </ModalButton>
-        <ModalButton
-          onPress={() => {
-            setIsModalVisible(!isModalVisible);
-            setIsDeleteModalVisible(!isDeleteModalVisible);
-          }}>
-          <CalloutTypo color={'black'}>삭제하기</CalloutTypo>
-        </ModalButton>
-      </ModalView>
-    </Modal>
+    <ModalView onPress={() => console.log('pressed')}>
+      <ModalButton first onPress={() => console.log('pressed')}>
+        <CalloutTypo color={'black'}>고정하기</CalloutTypo>
+      </ModalButton>
+      <ModalButton onPress={() => console.log('pressed')}>
+        <CalloutTypo color={'black'}>편집하기</CalloutTypo>
+      </ModalButton>
+      <ModalButton
+        onPress={() => {
+          console.log('pressed');
+        }}>
+        <CalloutTypo color={'black'}>삭제하기</CalloutTypo>
+      </ModalButton>
+    </ModalView>
   );
 };
